organize_code
"Removed unused imports, deleted unnecessary methods, and refactored code in various controllers and routes."
