+++ app/Http/Controllers/Api/V1/AuthController.php
@@ -9,7 +9,6 @@ use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 
-
 class AuthController extends BaseController
 {
     public function login(Request $request)
