+++ app/Http/Controllers/BaseController.php
@@ -2,8 +2,6 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
-
 class BaseController extends Controller
 {
     public function sendResponse($message, $result = [], $opt = [])
